backend/requirements.txt
Flask==2.0.1
Pillow==8.2.0
Flask-Cors==3.0.10
tensorflow==2.5.0
